requirements.txt
tda-api==1.5.2